refactor: move flow-runs inside a flows subdirectory
