+++ packages/backend/src/app/database/database-connection.ts
@@ -8,7 +8,7 @@ import { FlowVersionEntity } from '../flows/flow-version/flow-version-entity'
 import { FileEntity } from '../file/file.entity'
 import { StoreEntryEntity } from '../store-entry/store-entry-entity'
 import { InstanceEntity } from '../instance/instance.entity'
-import { FlowRunEntity } from '../flow-run/flow-run-entity'
+import { FlowRunEntity } from '../flows/flow-run/flow-run-entity'
 import { FlagEntity } from '../flags/flag.entity'
 import { system } from '../helper/system/system'
 import { SystemProp } from '../helper/system/system-prop'
