+++ packages/backend/src/app/flows/flow-run/flow-run-side-effects.ts
@@ -1,5 +1,5 @@
 import { FlowRun } from '@activepieces/shared'
-import { flowQueue } from '../workers/flow-worker/flow-queue'
+import { flowQueue } from '../../workers/flow-worker/flow-queue'
 
 type StartParams = {
     flowRun: FlowRun
