+++ packages/backend/src/app/workers/flow-worker/flow-queue-consumer.ts
@@ -1,7 +1,7 @@
 import { Worker } from 'bullmq'
 import { ActivepiecesError, ApId, ErrorCode, RunEnvironment, TriggerType } from '@activepieces/shared'
 import { createRedisClient } from '../../database/redis-connection'
-import { flowRunService } from '../../flow-run/flow-run-service'
+import { flowRunService } from '../../flows/flow-run/flow-run-service'
 import { triggerUtils } from '../../helper/trigger-utils'
 import { ONE_TIME_JOB_QUEUE, REPEATABLE_JOB_QUEUE } from './flow-queue'
 import { flowWorker } from './flow-worker'
