+++ packages/backend/src/app/workers/flow-worker/flow-worker.ts
@@ -17,7 +17,7 @@ import { Sandbox, sandboxManager } from '../sandbox'
 import { flowVersionService } from '../../flows/flow-version/flow-version.service'
 import { fileService } from '../../file/file.service'
 import { codeBuilder } from '../code-worker/code-builder'
-import { flowRunService } from '../../flow-run/flow-run-service'
+import { flowRunService } from '../../flows/flow-run/flow-run-service'
 import { OneTimeJobData } from './job-data'
 import { collectionService } from '../../collections/collection.service'
 import { engineHelper } from '../../helper/engine-helper'
