+++ packages/backend/src/main.ts
@@ -12,7 +12,7 @@ import { piecesController } from './app/pieces/pieces.controller'
 import { tokenVerifyMiddleware } from './app/authentication/token-verify-middleware'
 import { storeEntryModule } from './app/store-entry/store-entry.module'
 import { instanceModule } from './app/instance/instance.module'
-import { flowRunModule } from './app/flow-run/flow-run-module'
+import { flowRunModule } from './app/flows/flow-run/flow-run-module'
 import { flagModule } from './app/flags/flag.module'
 import { codeModule } from './app/workers/code-worker/code.module'
 import { flowWorkerModule } from './app/workers/flow-worker/flow-worker-module'
